requirements.txt
langchain==0.3.9
langchain_openai==0.2.11
python-dotenv==1.0.1
streamlit==1.39.0
streamlit_chat==0.1.1
